Renew the project as a starter template
I plan to use this repository as a starter template for new Git
repositories instead of having to install and maintain a set of npm
packages. I'll transfer the repository back to my own GitHub profile.

The easiest way to get started is by simply copying the repository
content, for example:

```shell
yarn dlx degit spdiswal/starter-web
```

This commit deletes the existing source code for the ESLint and Prettier
presets and replaces the monorepo project structure with an ordinary
project structure.

While I think AsciiDoc is a superior format to Markdown, Markdown has
better compatibility in npm and GitHub as well as in IDEs, which is why
this commit also switches to Markdown for documentation.